fix: git shows previous changes as my current ones bb-129
